# docs/swagger.yaml
basePath: https://api.web95.tech/api/v1
definitions:
  helpers.ErrorData:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  models.Countdown:
    properties:
      days:
        type: integer
      hours:
        type: integer
      minutes:
        type: integer
      seconds:
        type: integer
    type: object
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role_id:
        type: integer
    type: object
  models.CurrencyRatesPayload:
    properties:
      sar:
        type: number
      usd:
        type: number
    type: object
...
  models.HotelResponse:
    properties:
      city:
        type: string
      created_at:
        type: string
      distance:
        type: number
      distance_text:
        type: string
      guests:
        type: string
      id:
        type: integer
      meals:
        type: string
      name:
        type: string
      nights:
        type: integer
      stars:
        type: integer
      transfer:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
...
  models.News:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      media_type:
        type: string
      published_at:
        type: string
      reposts_count:
        type: integer
      slug:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок вместо preview_url"
        items:
          type: string
        type: array
      video_url:
        type: string
      views_count:
        type: integer
    type: object
...
  models.Trip:
    properties:
      active:
        type: boolean
      booking_deadline:
        type: string
      buys_count:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      departure_city:
        type: string
      description:
        type: string
      discount_percent:
        type: integer
      end_date:
        type: string
      final_price:
        type: number
      hotels:
        items:
          $ref: '#/definitions/models.TripHotelWithInfo'
        type: array
      id:
        type: integer
      main:
        type: boolean
      price:
        type: number
      season:
        type: string
      start_date:
        type: string
      title:
        type: string
      trip_type:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок вместо photo_url"
        items:
          type: string
        type: array
      views_count:
        type: integer
    type: object
...
  models.TripHotelWithInfo:
    properties:
      city:
        type: string
      distance:
        type: number
      hotel_id:
        type: integer
      meals:
        type: string
      name:
        type: string
      nights:
        type: integer
      rating:
        type: integer
    type: object
  models.TripOptionResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      unit:
        type: string
    type: object
  models.TripPageResponse:
    properties:
      countdown:
        $ref: '#/definitions/models.Countdown'
      currency_rates:
        $ref: '#/definitions/models.CurrencyRatesPayload'
      duration_days:
        type: integer
      hotels:
        items:
          $ref: '#/definitions/models.HotelResponse'
        type: array
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
      options:
        description: "\U0001F539 новые доп.опции"
        items:
          $ref: '#/definitions/models.TripOptionResponse'
        type: array
      popular_trips:
        items:
          $ref: '#/definitions/models.Trip'
        type: array
      reviews:
        $ref: '#/definitions/models.TripPageReviews'
      routes:
        $ref: '#/definitions/models.TripRouteResponse'
      trip:
        $ref: '#/definitions/models.Trip'
    type: object
  models.TripPageReviews:
    properties:
      items:
        items:
          $ref: '#/definitions/models.TripReview'
        type: array
      total:
        type: integer
    type: object
  models.TripReview:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      trip_id:
        type: integer
      user_name:
        type: string
    type: object
...
  models.TripRouteCitiesResponse:
    properties:
      route_cities:
        additionalProperties:
          $ref: '#/definitions/models.TripRouteCity'
        type: object
    type: object
  models.TripRouteCity:
    properties:
      city:
        type: string
      duration:
        type: string
      stop_time:
        type: string
      transport:
        type: string
    required:
    - city
    type: object
...
  models.TripRouteResponse:
    properties:
      route:
        items:
          $ref: '#/definitions/models.TripRouteSegment'
        type: array
      total_duration:
        type: string
    type: object
  models.TripRouteSegment:
    properties:
      city:
        type: string
      duration:
        type: string
      stop_time:
        type: string
      transport:
        type: string
    type: object
...
  models.TripWithRelations:
    properties:
      hotels:
        items:
          $ref: '#/definitions/models.HotelResponse'
        type: array
      routes:
        $ref: '#/definitions/models.TripRouteCitiesResponse'
      trip:
        $ref: '#/definitions/models.Trip'
    type: object
...
  models.UpdateUserRequest:
    properties:
      full_name:
        type: string
      role_id:
        type: integer
    type: object
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Travel project
  title: Travel API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Не удалось получить список пользователей
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Получить всех пользователей
      tags:
      - Admin — Users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректное тело запроса
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при создании пользователя
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Создать пользователя
      tags:
      - Admin — Users
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при удалении пользователя
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Удалить пользователя
      tags:
      - Admin — Users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при получении пользователя
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Получить пользователя по ID
      tags:
      - Admin — Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректное тело запроса
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при обновлении пользователя
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Обновить данные пользователя
      tags:
      - Admin — Users
...
  /trips/{id}/page:
    get:
      description: Полный набор данных для страницы тура (тур, отели, отзывы, популярные
        туры, новости, курсы, countdown)
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Поиск по названию тура
        in: query
        name: title
        type: string
      - description: Город вылета
        in: query
        name: departure_city
        type: string
      - description: Тип тура
        in: query
        name: trip_type
        type: string
      - description: Сезон
        in: query
        name: season
        type: string
      - description: Город в маршруте
        in: query
        name: route_city
        type: string
      - description: Статус тура
        in: query
        name: active
        type: boolean
      - description: Дата начала с (YYYY-MM-DD)
        in: query
        name: start_after
        type: string
      - description: Дата окончания до (YYYY-MM-DD)
        in: query
        name: end_before
        type: string
      - description: Лимит (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TripPageResponse'
        "404":
          description: Тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Trip page data
      tags:
      - Public — Trips
  /trips/{id}/relations:
    get:
      description: Возвращает тур вместе с отелями и маршрутом
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TripWithRelations'
        "404":
          description: Тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get trip with hotels and routes
      tags:
      - Public — Trips
...
  /trips/full:
    get:
      description: Возвращает список всех туров вместе с отелями, маршрутами и опциями
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TripPageResponse'
            type: array
        "500":
          description: Ошибка при получении туров
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: All trips with full data
      tags:
      - Public — Trips
...
  /trips/relations:
    get:
      description: Возвращает все туры вместе с отелями и маршрутами
      parameters:
      - description: Поиск по названию тура
        in: query
        name: title
        type: string
      - description: Город вылета
        in: query
        name: departure_city
        type: string
      - description: Тип тура
        in: query
        name: trip_type
        type: string
      - description: Сезон
        in: query
        name: season
        type: string
      - description: Город в маршруте
        in: query
        name: route_city
        type: string
      - description: Статус тура
        in: query
        name: active
        type: boolean
      - description: Дата начала с (YYYY-MM-DD)
        in: query
        name: start_after
        type: string
      - description: Дата окончания до (YYYY-MM-DD)
        in: query
        name: end_before
        type: string
      - description: Лимит (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TripWithRelations'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: List all trips with hotels and routes
      tags:
      - Public — Trips
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"